Move mutex into subdirectories
