+++ Makefile.am
@@ -23,8 +23,11 @@ AM_CFLAGS += -DKERNAUX_PRIVATE_NO
 
 lib_LTLIBRARIES = libkernaux.la
 
-libkernaux_la_SOURCES = src/libc.h src/assert.c src/mutex.c
 libkernaux_la_LIBADD =
+libkernaux_la_SOURCES = \
+	src/assert.c \
+	src/libc.h \
+	src/generic/mutex.c
 
 if ASM_I386
 libkernaux_la_SOURCES += src/asm/i386.S
