+++ README.md
@@ -44,11 +44,12 @@ zero). Work-in-progress APIs can change at any time.
     * [Example: Assert](/examples/assert.c)
     * [Example: Panic](/examples/panic.c)
   * Stack trace *(planned)*
-  * [Mutex](/include/kernaux/mutex.h) (*non-breaking since* **?.?.?**)
   * [Input/output](/include/kernaux/io.h) (*work in progress*)
   * Architecture-specific code (*work in progress*)
     * [Declarations](/include/kernaux/arch/)
     * [Functions](/include/kernaux/asm/)
+* Generic types
+  * [Mutex](/include/kernaux/generic/mutex.h) (*non-breaking since* **?.?.?**)
 * Device drivers (for debugging only)
   * [Serial console](/include/kernaux/console.h) (*work in progress*)
   * [Framebuffer](/include/kernaux/framebuffer.h) (*planned*)
